Ux improvement

+++ FlavorFlash/Services/FBDataService/AuthenticationManager.swift
@@ -45,9 +45,9 @@ final class AuthenticationManager {
 	// MARK: - helper getter
 	func getAuthenticatedUser() throws -> AuthDataResultModel {
 		guard let user = Auth.auth().currentUser else {
-			throw URLError(.badServerResponse)
+			throw FBAuthError.userNotLoggedIn
 		}
-		debugPrint("current loggedin user: \(user.uid)")
+		logger.info("Current Logged in UserID: \(user.uid)")
 
 		return AuthDataResultModel(user: user)
 	}
@@ -73,7 +73,10 @@ final class AuthenticationManager {
 	}
 
 	func signOut() throws {
-		debugPrint("\(Auth.auth().currentUser?.displayName) signing out")
+		guard let user = Auth.auth().currentUser else {
+			throw FBAuthError.userNotLoggedIn
+		}
+		logger.info("\(user.uid) signing out")
 		try Auth.auth().signOut()
 	}
 
@@ -90,8 +93,11 @@ final class AuthenticationManager {
 	}
 
 	func deleteAccount() {
-		debugPrint("delete currentUser : \(Auth.auth().currentUser)")
+		guard let user = Auth.auth().currentUser else {
+			return
+		}
 		Auth.auth().currentUser?.delete()
+		logger.info("\(user.uid) deleted account")
 	}
 }
 
@@ -127,7 +133,12 @@ extension AuthenticationManager {
 extension AuthenticationManager {
 
 	func sendTokenToServer(token: String?) {
-		var deviceToken: [String: Any] = [
+		guard let token else {
+			logger.info("FCM Token is nil")
+			return
+		}
+
+		let deviceToken: [String: Any] = [
 			"token": token,
 			"timestamp": FieldValue.serverTimestamp()
 		]
